Switch el-get from 3.stable to master branch
If you're pulling down this commit, you will probably have to nuke your
el-get directory with all installed packages and init/update the
submodules again. A lot of things have changed in el-get since 3.1 was
released.

I made the move to be able to use a couple of new features for
package/recipe definitions. And after the initial reinstall of all
packages everything worked like before :)
Switch from 3.stable to master branch of el-get

If you're pulling down this commit, you will probably have to nuke your
el-get directory with all installed packages and init/update the
submodules again. A lot of things have changed in el-get since 3.1 was
released.

I made the move to be able to use a couple of new features for
package/recipe definitions. And after the initial reinstall of all
packages everything worked like before :)