Rework package archive setup, preferring melpa-stable over melpa
Changed up package-archive setup again, this time we just hard-code all
the archive sources, rather than do fancy detection of if we can use
HTTPS or not.

Also we're throwing MELPA Stable into the mix, and giving it higher
priority than regular MELPA. The idea is to try to typically install
stable packages, but fall-back onto the regular MELPA source if there is
no stable version of it.

First impressions with a clean re-install of all packages is
promising. Time will tell if I have any issues.