Ensure module names cannot collide with core setup files
The core setup files for Emacs Siren which lives in the core directory
followed a `siren-*.el` naming convention, which is the same as the
naming convention for modules.

This means that the `modules/core/siren-packages.el` module for adding
packages for Emacs package development, was not being loaded due to it's
name conflicting with `core/siren-packages.el` which sets up and
configures the packaging system.

So all files under the root `core` directory now follow a
`siren-core-*.el` naming scheme, meaning modules should no longer
conflict with core files.