chore(core): Make Emacs play nice when run in a terminal
Ever since I started using persp-mode, launching Emacs in a terminal has
made it restore the same set of open files as in my main GUI instance of
Emacs. This hasn't been a big deal, but still somewhat annoying.

I've finally gotten around to disabling persp-mode's auto save/load
feature when running Emacs in a terminal.

And for good measure, I also set the theme to be doom-vibrant in the
terminal too. My old tomorrow-night-paradise is way out of date, and
nowhere near as good as doom-vibrant.