fix(workspaces): Don't restore magit buffers on startup
Persp-mode doesn't quite manage to restore magit buffers correctly, so
I've opted to have it ignore any and all magit buffers when performing
it's auto-save.

It is possible to configure it to know how to save and load the
magit-status buffer for example, but the snippet for it from
persp-mode's readme doesn't fully restore the buffer as is. And magit
has many different kinds of buffers, so I've simply opted for having all
ignored.