fix(vc): Disable magit-todos package
It seems to cause very long freezes within magit itself, especially when
trying to commit. It seems to that the git diff command that's run in
the background hangs and/or takes a very long time to complete for some
reason.

Manually killing those git processes let's magit wake up instantly
again, in htop they're listed as:

    git --no-pager diff --no-color -U0 <Git SHA of HEAD>