feat(persp-mode): Enable switching to most recent persp
Keep a list of most recently visited persps whenever you switch, and
enable switching back and forth between the current and most recently
visited persp.

Also when killing current persp, switch to the most recently visited
persp, instead of back to the default/nil persp.

Minor bug remaining is that the when killing current persp, the echo
area shows all persps like usual when switching, but it includes the
persp that was just killed, despite it already being killed.