fix(workspaces): do not auto-load "default" desktop on startup
If you launch Emacs by giving it a file to open, the auto-loading of the
default desktop would remove that file, as it is not likely to be part
of the default desktop.

Instead desktop management is now disabled until you specifically load a
desktop with `desktop+-load-or-create` (bound to `C-z C-z C-s`).