chore(startup): stop using emacs-startup-hook to trigger package load
Packages that need to be loaded on emacs startup, should just be loaded
through use-package in a non-deferred manner.

It makes no real difference to startup, loading the packages either
slows down emacs before "startup" is complete, or right after it
completes. End result is that Emacs is unresponsive for basically the
same amount of time regardless.