feat(version-control): tweak diff-hl style
Change style back to a plain vertical line, but slightly narrower than
last time which I ended up not liking, while also aligning it to the
right hand side of the left fringe, so there's a gap between the window
border and the diff-hl status line. Also tweak the colors a bit.

I feel this works better, and keeps things nice and clean.