fix(dired): prevent dired from prompting about removing dired buffers
When deleting a file in Dired, it has recently started prompting about
removing the dired buffer along with file buffer. This is rather
annoying, as I almost never want to close the dired instance I have open
when simply deleting a file from a folder.

So I've overridden the relevant function, and added an option
specifically to enable/disable the prompt about removing the dired
buffer.