fix(language/terraform): disable lsp-mode in terraform files
The terraform language server is not very good, very rarely provides any
completions, and when it doesn't it overall is just slow and laggy.

Especially so as lsp-mode is set to syntax highlight with the language
server, causing syntax highlighting to often be about 5 seconds behind
of whatever you're typing.