chore(themes/doom-overrides): use doom-themes-set-faces helper
The doom-themes-set-faces essentially just lets you apply a doom-themes
faces spec against any theme you want.

Since we're defining a theme meant to override and extend various faces
from doom-themes, this makes it easier to do so, but also enables the
use of targeting light and dark theme variants with &light and &dark.