feat(completion/copilot): add accept completion "mods" concept
Essentially, if accepted completion string matches a given regexp
pattern, text can be inserted both before and after point.

This is useful in certain situations where Copilot suggests the opening
line to a if statement, for loop, etc., which leads to unbalanced curly
brackets. This can help reduce the annoyance of unbalanced brackets.