feat(theme/override): invert tab-bar active/inactive background colors
Previously the active tab was the same background color as normal
buffers, while the rest of the tab-bar had an alternative background
color.

I feel that it visually fits better when these are swapped, so tab-bar
overall has the same background color as normal buffers, and the active
tab has an alternative background color.