feat(keybindings): add cmd+shift+p and cmd+/ keybindings from VSCode
Makes cmd+shift+p open the command palette and cmd+/ toggle comments, as
they do in VSCode.

This should ease my recent usage of Cursor/VSCode alongside Emacs, as
certain keybindings I have in Emacs cannot be setup in VSCode to work in
all contexts. Hence I've had to get used to cmd+shift+p instead of C-x
C-m, as the later does not want to play ball when a terminal is active
in VSCode.